dev : add CSE IP Address UI moved resourceType and attributes to seperate file
