+++ oneM2M_Data_Model_Design_Tool/src/App.vue
@@ -5,10 +5,13 @@
   <div class="configure">
     <div class="box">
       <div class="key">CSE IP address</div>
-      <input type="text" :model="targetIP" placeholder="http://127.0.0.1:3000/TinyIoT" />
+      <input 
+      v-model="targetIP" 
+      placeholder="http://127.0.0.1:3000/TinyIoT" 
+      />
     </div>
     <div>
-        <div class="btn button">Load</div>
+        <div class="btn button" @click="loadFromRemoteCSE">Load</div>
     </div>
   </div>
   <div class="body">
@@ -136,21 +139,8 @@ import draggable from "vuedraggable";
 import nestedDraggable from "@/components/infra/nested.vue";
 import setAttrs from "@/components/setAttrs.vue";
 import navBar from "@/components/navBar.vue";
+import { resourceType as RT } from "./components/attributes";
 import get_jsonfile from "@/components/json-parser.js";
-
-const RT_CSE = 5;
-const RT_ACP = 1;
-const RT_AE = 2;
-const RT_CNT = 3;
-const RT_GRP = 9;
-const RT_SUB = 23;
-const RT_FCNT = 7;
-const RT_TS = 8;
-const RT_TSI = 9;
-const RT_TSR = 10;
-const RT_MGMTOBJ = 11;
-const RT_NODE = 14;
-
 
 export default {
   name: "App",
@@ -168,7 +158,7 @@ export default {
       cse1: [
       {
           name: "CSE1",
-          ty: RT_CSE,
+          ty: RT.CSE,
           tasks: [
           ],
           attrs:{
@@ -177,11 +167,11 @@ export default {
         }
       ],
       resources: [
-          { name: "AE", ty: RT_AE },
-          { name: "CNT", ty: RT_CNT },
-          { name: "ACP", ty: RT_ACP },
-          { name: "GRP", ty: RT_GRP },
-          { name: "SUB", ty: RT_SUB },
+          { name: "AE", ty: RT.RT_AE },
+          { name: "CNT", ty: RT.RT_CNT },
+          { name: "ACP", ty: RT.RT_ACP },
+          { name: "GRP", ty: RT.RT_GRP },
+          { name: "SUB", ty: RT.RT_SUB },
           // { name: "FCNT", ty: RT_FCNT },
       ]
       ,
@@ -189,7 +179,7 @@ export default {
       attrSettingModified: false,
       isDragging: false,
       selectedElement: {},
-      targetIP:""
+      targetIP: ""
     };
 
   },
@@ -197,6 +187,9 @@ export default {
     const cse = JSON.parse(sessionStorage.getItem("CSE1"));
     //get_jsonfile(cse);
     if (cse!=undefined) this.cse1 = cse;
+  },
+  mounted(){
+    this.targetIP = sessionStorage.getItem('targetIP');
   },
   methods: {
     saveResourceTree(){
@@ -208,8 +201,6 @@ export default {
       this.create_oneM2M_resource();
      // console.log("create finish");
     },
-    
-
     setAttributes(element){
       this.selectedElement = element;
       this.attrSettingModified = false;
@@ -257,6 +248,21 @@ export default {
       };
       fileInput.click();
     },
+    loadFromRemoteCSE(){
+      console.log("loadFromRemoteCSE");
+      console.log(this.targetIP);
+      sessionStorage.setItem('targetIP', this.targetIP);
+      if(this.targetIP === ""){
+        alert("Please input CSE IP address");
+        return;
+      }
+      const url = this.targetIP;
+      const protocol = url.split(':')[0];
+      const ip = url.split(':')[1].replace('//','');
+      const port = url.split(':')[2].split('/')[0];
+      const path = url.split(':')[2].split('/').slice(1).join('/');
+      console.log(protocol, ip, port, path);
+    },
     loadFromSessionStorage() {
       const data = sessionStorage.getItem('CSE1');
       if (data) {
@@ -291,7 +297,7 @@ export default {
           }
         }
         const attribute = task.attrs;
-        if(task.ty == RT_AE){ /* AE */
+        if(task.ty == RT.AE){ /* AE */
           if(
             (typeof attribute.api == "undefined" || typeof attribute.rr == "undefined" || typeof attribute.srv == "undefined") ||             // Mandatory Attribute
             (typeof attribute.rn !== "undefined" && !/^[a-zA-Z0-9\-._]*$/.test(attribute.rn)) ||                                              // resourceName
@@ -317,7 +323,7 @@ export default {
             }
           }
         }
-        if(task.ty == RT_CNT){ /* CNT */
+        if(task.ty == RT.CNT){ /* CNT */
           if(
             (typeof attribute.lbl !== "undefined" && !/^[a-zA-Z0-9:]*$/.test(attribute.lbl)) ||                                               // labels
             (typeof attribute.acpi !== "undefined" && typeof attribute.acpi !== 'string') ||                                                  // accessControlPolicyIDs
@@ -333,7 +339,7 @@ export default {
             return false;
           }
         }        
-        if(task.ty == RT_SUB){ /* SUB */
+        if(task.ty == RT.SUB){ /* SUB */
           if(
             (typeof attribute.nu == "undefined") ||                                                                                           // Mandatory Attribute
             (typeof attribute.lbl !== "undefined" && !/^[a-zA-Z0-9:]*$/.test(attribute.lbl)) ||                                               // labels
@@ -349,7 +355,7 @@ export default {
             return false;
           }
         }
-        if(task.ty == RT_GRP){ /* GRP */
+        if(task.ty == RT.GRP){ /* GRP */
           if(
             (typeof attribute.mnm == "undefined" || typeof attribute.mid == "undefined") ||                                                   // Mandatory Attribute
             (typeof attribute.lbl !== "undefined" && !/^[a-zA-Z0-9:]*$/.test(attribute.lbl)) ||                                               // labels
